added template info transferring with additional functionalities

+++ resources/views/images/approved.blade.php
@@ -2,13 +2,40 @@
 
 
 @section('content')
+  <link rel="stylesheet" type="text/css" href="https://cdnjs.cloudflare.com/ajax/libs/bootstrap-multiselect/0.9.15/css/bootstrap-multiselect.css">
+
 <div class="row">
   <div class="col-12 margin-tb mb-3">
     <h2 class="page-heading">Approved Images</h2>
 
+    <form action="{{ route('image.grid.approved') }}" method="GET" class="form-inline align-items-start">
+      <div class="form-group mr-3 mb-3">
+        {!! $category_selection !!}
+      </div>
+
+      <div class="form-group mr-3">
+        <select class="form-control select-multiple" name="brand[]" multiple>
+          <optgroup label="Brands">
+            @foreach ($brands as $key => $name)
+              <option value="{{ $key }}" {{ isset($brand) && $brand == $key ? 'selected' : '' }}>{{ $name }}</option>
+            @endforeach
+        </optgroup>
+        </select>
+      </div>
+
+      <div class="form-group mr-3">
+        <strong class="mr-3">Price</strong>
+        <input type="text" name="price" data-provide="slider" data-slider-min="0" data-slider-max="10000000" data-slider-step="10" data-slider-value="[{{ isset($price) ? $price[0] : '0' }},{{ isset($price) ? $price[1] : '10000000' }}]"/>
+      </div>
+
+      <button type="submit" class="btn btn-image"><img src="/images/filter.png" /></button>
+    </form>
+
     @can('social-create')
       <div class="pull-right">
-        <a href class="btn btn-secondary" data-toggle="modal" data-target="#imageModal">Upload Templates</a>
+        {{-- <a href class="btn btn-secondary" data-toggle="modal" data-target="#imageModal">Upload Templates</a> --}}
+        <a href class="btn btn-secondary" id="toggleButton" data-toggle="modal" data-target="#imageModal" style="display: none;">Upload Templates</a>
+        <a href class="btn btn-secondary select-image">Upload Templates</a>
       </div>
 
       <div id="imageModal" class="modal fade" role="dialog">
@@ -24,10 +51,11 @@
             <form action="{{ route('image.grid.store') }}" method="POST" enctype="multipart/form-data">
               @csrf
               <input type="hidden" name="status" value="2">
+              <input type="hidden" name="image_id" value="">
 
               <div class="modal-body">
                 <div class="form-group">
-                     <input type="file" name="images[]" multiple required />
+                     <input type="file" name="images[]" required />
                      @if ($errors->has('images'))
                          <div class="alert alert-danger">{{$errors->first('images')}}</div>
                      @endif
@@ -58,13 +86,16 @@
   @foreach ($images as $image)
   <div class="col-md-3 col-xs-6 text-center mb-5">
     <img src="{{ $image->filename ? (asset('uploads/social-media') . '/' . $image->filename) : ($image->getMedia(config('constants.media_tags'))->first() ? $image->getMedia(config('constants.media_tags'))->first()->getUrl() : '') }}" class="img-responsive grid-image" alt="" />
-
+    <input type="checkbox" class="form-control image-selection" value="{{ $image->id }}" style="display: none;">
     <a class="btn btn-image" href="{{ route('image.grid.show',$image->id) }}"><img src="/images/view.png" /></a>
-    {{-- <a class="btn btn-image" href="{{ route('image.grid.edit',$image->id) }}"><img src="/images/edit.png" /></a>
 
-    {!! Form::open(['method' => 'DELETE','route' => ['image.grid.delete', $image->id],'style'=>'display:inline']) !!}
-      <button type="submit" class="btn btn-image"><img src="/images/delete.png" /></button>
-    {!! Form::close() !!} --}}
+    @can ('social-create')
+      <a class="btn btn-image" href="{{ route('image.grid.edit',$image->id) }}"><img src="/images/edit.png" /></a>
+
+      {!! Form::open(['method' => 'DELETE','route' => ['image.grid.delete', $image->id],'style'=>'display:inline']) !!}
+        <button type="submit" class="btn btn-image"><img src="/images/delete.png" /></button>
+      {!! Form::close() !!}
+    @endcan}
 
     @if (isset($image->approved_user))
       <span>Approved by {{ App\User::find($image->approved_user)->name}} on {{ Carbon\Carbon::parse($image->approved_date)->format('d-m') }}</span>
@@ -75,17 +106,25 @@
 
 {!! $images->appends(Request::except('page'))->links() !!}
 
+  <script src="https://cdnjs.cloudflare.com/ajax/libs/bootstrap-multiselect/0.9.15/js/bootstrap-multiselect.min.js"></script>
+  <script type="text/javascript">
+  $(document).ready(function() {
+     $(".select-multiple").multiselect();
 
-<script>
-  // var searchSuggestions = ;
-  var image_array = [];
+     $('.select-image').on('click', function(e) {
+       e.preventDefault();
 
-  $('#product-search').autocomplete({
-    source: function(request, response) {
-      var results = $.ui.autocomplete.filter(searchSuggestions, request.term);
+       $('.image-selection').show();
+     });
 
-      response(results.slice(0, 10));
-    }
+     $(document).on('click', '.image-selection', function() {
+       $('.image-selection').prop('checked', false);
+
+       $(this).prop('checked', true);
+       $('input[name="image_id"]').val($(this).val());
+       jQuery.noConflict();
+       $('#imageModal').modal('toggle')
+     });
   });
-</script>
+  </script>
 @endsection
