+++ resources/views/images/final.blade.php
@@ -2,9 +2,34 @@
 
 
 @section('content')
+  <link rel="stylesheet" type="text/css" href="https://cdnjs.cloudflare.com/ajax/libs/bootstrap-multiselect/0.9.15/css/bootstrap-multiselect.css">
+
 <div class="row">
   <div class="col-12 margin-tb mb-3">
     <h2 class="page-heading">Final Approval</h2>
+
+    <form action="{{ route('image.grid.final.approval') }}" method="GET" class="form-inline align-items-start">
+      <div class="form-group mr-3 mb-3">
+        {!! $category_selection !!}
+      </div>
+
+      <div class="form-group mr-3">
+        <select class="form-control select-multiple" name="brand[]" multiple>
+          <optgroup label="Brands">
+            @foreach ($brands as $key => $name)
+              <option value="{{ $key }}" {{ isset($brand) && $brand == $key ? 'selected' : '' }}>{{ $name }}</option>
+            @endforeach
+        </optgroup>
+        </select>
+      </div>
+
+      <div class="form-group mr-3">
+        <strong class="mr-3">Price</strong>
+        <input type="text" name="price" data-provide="slider" data-slider-min="0" data-slider-max="10000000" data-slider-step="10" data-slider-value="[{{ isset($price) ? $price[0] : '0' }},{{ isset($price) ? $price[1] : '10000000' }}]"/>
+      </div>
+
+      <button type="submit" class="btn btn-image"><img src="/images/filter.png" /></button>
+    </form>
   </div>
 </div>
 
@@ -20,11 +45,15 @@
   <div class="col-md-3 col-xs-6 text-center mb-5">
     <img src="{{ $image->filename ? (asset('uploads/social-media') . '/' . $image->filename) : ($image->getMedia(config('constants.media_tags'))->first() ? $image->getMedia(config('constants.media_tags'))->first()->getUrl() : '') }}" class="img-responsive grid-image" alt="" />
 
-    {{-- <a class="btn btn-image" href="{{ route('image.grid.edit',$image->id) }}"><img src="/images/edit.png" /></a>
+    <a class="btn btn-image" href="{{ route('image.grid.show',$image->id) }}"><img src="/images/view.png" /></a>
 
-    {!! Form::open(['method' => 'DELETE','route' => ['image.grid.delete', $image->id],'style'=>'display:inline']) !!}
-      <button type="submit" class="btn btn-image"><img src="/images/delete.png" /></button>
-    {!! Form::close() !!} --}}
+    @can ('social-create')
+      <a class="btn btn-image" href="{{ route('image.grid.edit',$image->id) }}"><img src="/images/edit.png" /></a>
+
+      {!! Form::open(['method' => 'DELETE','route' => ['image.grid.delete', $image->id],'style'=>'display:inline']) !!}
+        <button type="submit" class="btn btn-image"><img src="/images/delete.png" /></button>
+      {!! Form::close() !!}
+    @endcan
 
     @if (isset($image->approved_user))
       <span>Approved by {{ App\User::find($image->approved_user)->name}} on {{ Carbon\Carbon::parse($image->approved_date)->format('d-m') }}</span>
@@ -42,17 +71,10 @@
 
 {!! $images->appends(Request::except('page'))->links() !!}
 
-
-{{-- <script>
-  // var searchSuggestions = ;
-  var image_array = [];
-
-  $('#product-search').autocomplete({
-    source: function(request, response) {
-      var results = $.ui.autocomplete.filter(searchSuggestions, request.term);
-
-      response(results.slice(0, 10));
-    }
+  <script src="https://cdnjs.cloudflare.com/ajax/libs/bootstrap-multiselect/0.9.15/js/bootstrap-multiselect.min.js"></script>
+  <script type="text/javascript">
+  $(document).ready(function() {
+     $(".select-multiple").multiselect();
   });
-</script> --}}
+  </script>
 @endsection
